fix some more warnings from msvc

+++ protocol/server/main.c
@@ -183,12 +183,12 @@ void create_and_send_tcp_wmsg(WhistServerMessageType message_type, char* payload
             break;
         }
         case SMESSAGE_FILE_METADATA: {
-            data_size = ((FileMetadata*)payload)->filename_len;
+            data_size = (int)((FileMetadata*)payload)->filename_len;
             type_size = sizeof(FileMetadata);
             break;
         }
         case SMESSAGE_FILE_DATA: {
-            data_size = ((FileData*)payload)->size;
+            data_size = (int)((FileData*)payload)->size;
             type_size = sizeof(FileData);
             break;
         }
